requirements.txt
streamlit
scikit-learn
pandas
numpy
joblib
newspaper3k
